requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
tensorflow-cpu==2.13.0
typing_extensions==4.5.0
altair==4.2.2
